// src/BookBarter.React/src/api/generated/models/update-publisher-command.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BookBarter.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface UpdatePublisherCommand
 */
export interface UpdatePublisherCommand {
  /**
   *
   * @type {number}
   * @memberof UpdatePublisherCommand
   */
  id?: number;
  /**
   *
   * @type {string}
   * @memberof UpdatePublisherCommand
   */
  name?: string | null;
}